moved InputType enum into rate-based converter
