+++ ml_genn/converters/__init__.py
@@ -1,2 +1,2 @@
 from ml_genn.converters.few_spike import FewSpike
-from ml_genn.converters.rate_based import RateBased
+from ml_genn.converters.rate_based import InputType, RateBased
