+++ ml_genn/layers/__init__.py
@@ -1,5 +1,3 @@
-
-from ml_genn.layers.enum import InputType
 from ml_genn.layers.enum import ConnectivityType
 from ml_genn.layers.enum import PadMode
